fix: url pattern to follow standard
