+++ django_sso/users/api/views.py
@@ -1,71 +1,32 @@
 import json
 import secrets
-from urllib.parse import urlencode
 
 import jwt
 
 # django
 from django.conf import settings
 from django.contrib.auth import get_user_model
-from django.contrib.auth.forms import AuthenticationForm
 from django.contrib.auth.hashers import check_password
-from django.contrib.auth.mixins import LoginRequiredMixin
-from django.contrib.auth.views import LoginView as DJLoginView
 from django.core.cache import cache
 from django.http import JsonResponse
-from django.shortcuts import redirect
-from django.urls import reverse
-from django.utils.http import url_has_allowed_host_and_scheme
 from django.utils.timezone import now
-from django.views import View
-from django.views.generic.edit import FormView
 
 # rest framework
 from rest_framework import permissions, status
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from django_sso.utils.string import normalize_uri
-
-from .forms import RegisterForm
-from .models import Application
-
 # local
-from .serializers import TokenRequestSerializer
+# model
+from django_sso.users.models import Application
+
+# serializer
+from django_sso.users.serializers import TokenRequestSerializer
 
 # utils
-from .utils.auth import create_and_cache_auth_code
+from django_sso.utils.string import normalize_uri
 
 User = get_user_model()
-
-
-class AuthorizeView(LoginRequiredMixin, View):
-    login_url = "/api/users/login/"
-
-    def get(self, request, *args, **kwargs):
-        client_id = request.GET.get("client_id")
-        redirect_uri = request.GET.get("redirect_uri")
-        state = request.GET.get("state", "")
-        scope = request.GET.get("scope", "openid")
-        requested_scopes = scope.split()
-
-        client = Application.objects.filter(client_id=client_id, is_active=True).first()
-        if not client:
-            return redirect("/error?error=invalid_client")
-
-        allowed_scopes = set(client.get_allowed_scopes())
-        granted_scopes = [s for s in requested_scopes if s in allowed_scopes]
-
-        if normalize_uri(redirect_uri) not in client.get_redirect_uris():
-            return redirect("/error?error=invalid_redirect_uri")
-
-        if not granted_scopes:
-            return redirect("/error?error=invalid_scope")
-
-        auth_code = create_and_cache_auth_code(request.user, client, redirect_uri, granted_scopes)
-
-        query = urlencode({"code": auth_code, "state": state})
-        return redirect(f"{redirect_uri}?{query}")
 
 
 class TokenView(APIView):
@@ -204,7 +165,9 @@ class UserInfoView(APIView):
 
 
 class RefreshTokenView(APIView):
-    permission_classes = [permissions.AllowAny]
+    permission_classes = [
+        permissions.AllowAny,
+    ]
 
     def post(self, request, *args, **kwargs):
         refresh_token = request.data.get("refresh_token")
@@ -280,7 +243,9 @@ class RefreshTokenView(APIView):
 
 
 class LogoutView(APIView):
-    permission_classes = [permissions.AllowAny]
+    permission_classes = [
+        permissions.AllowAny,
+    ]
 
     def post(self, request):
         token = self._get_token_from_request(request)
@@ -306,44 +271,3 @@ class LogoutView(APIView):
         if auth_header.startswith("Bearer "):
             return auth_header.split("Bearer ")[1]
         return None
-
-
-class LoginView(DJLoginView):
-    template_name = "users/login.html"
-    authentication_form = AuthenticationForm
-    redirect_authenticated_user = True
-
-    def get_success_url(self):
-        redirect_to = self.request.GET.get("next")
-        if redirect_to and url_has_allowed_host_and_scheme(redirect_to, self.request.get_host()):
-            return redirect_to
-
-        return settings.LOGIN_REDIRECT_URL
-
-
-class RegisterView(FormView):
-    template_name = "users/register.html"
-    form_class = RegisterForm
-
-    def form_valid(self, form):
-        form.save()
-        next_url = self.request.POST.get("next", "")
-        login_url = reverse("accounts:login")
-        if next_url and url_has_allowed_host_and_scheme(next_url, self.request.get_host()):
-            login_url += f"?next={next_url}"
-        return redirect(login_url)
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["next"] = self.request.GET.get("next", "")
-        return context
-
-    def get_initial(self):
-        initial = super().get_initial()
-        initial["next"] = self.request.GET.get("next")
-        return initial
-
-    def get_form_kwargs(self):
-        kwargs = super().get_form_kwargs()
-        kwargs.update({"files": self.request.FILES})
-        return kwargs
