+++ django_sso/users/urls.py
@@ -1,16 +1,8 @@
-from django.urls import path
-
-from .views import AuthorizeView, LoginView, LogoutView, RefreshTokenView, RegisterView, TokenView, UserInfoView
+from django.urls import include, path
 
 app_name = "accounts"
 
 urlpatterns = [
-    path("logout/", LogoutView.as_view(), name="logout"),
-    path("login/", LoginView.as_view(), name="login"),
-    path("register/", RegisterView.as_view(), name="register"),
-    # token related
-    path("authorize/", AuthorizeView.as_view(), name="authorize"),
-    path("token/", TokenView.as_view(), name="token"),
-    path("token/refresh/", RefreshTokenView.as_view(), name="token"),
-    path("userinfo/", UserInfoView.as_view(), name="userinfo"),
+    path("users/", include(("django_sso.users.web.urls", "accounts"), namespace="web")),
+    path("api/users/", include(("django_sso.users.api.urls", "accounts"), namespace="api")),
 ]
